Mod05 - Capstone  (Gestor de Biblioteca)/project/requirements.txt
Flask==3.0.0
flask-marshmallow==1.2.1
marshmallow==3.20.1
Werkzeug==3.0.0
mysql-connector-python==8.2.0
requests==2.31.0